+++ example/lib/main.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'package:monobank_api/monobank_api.dart';
 
 // Connecting emoji dataset for MCC (transaction category)
@@ -10,17 +11,17 @@ void statement() async {
   // Create client
   final client = MonoAPI('token');
 
-// Request client
+  // Request client
   final res = await client.clientInfo();
 
-// Get first account
+  // Get first account
   final account = res.accounts[0];
 
   // Get statement list for last 3 months
   final statement = account.statement(
       DateTime.now().subtract(Duration(days: 31 * 3)), DateTime.now());
 
-// For each statement item
+  // For each statement item
   await for (final item in statement.list(reverse: true)) {
     // Output string representation
     print('${item.mcc.emoji} $item (${item.operationAmount.currency.name})');
@@ -55,6 +56,7 @@ void money() {
     Currency.code('USD'),
     1.165,
     1.182,
+    DateTime.now(),
     rounding: MoneyRounding.floor,
   );
 
@@ -78,4 +80,61 @@ void currency() async {
 
   // Printing
   print(result);
+}
+
+void webhook() async {
+  const webhookData = '''
+{
+  "type": "StatementItem",
+  "data": {
+    "account": "q5MA8eamezlw-SQjcddOmQ",
+    "statementItem": {
+      "id": "ZuHWzqkKGVo=",
+      "time": 1554466347,
+      "description": "Покупка щастя",
+      "mcc": 7997,
+      "originalMcc": 7997,
+      "hold": false,
+      "amount": -95000,
+      "operationAmount": -95000,
+      "currencyCode": 980,
+      "commissionRate": 0,
+      "cashbackAmount": 19000,
+      "balance": 10050000,
+      "comment": "За каву",
+      "receiptId": "XXXX-XXXX-XXXX-XXXX",
+      "invoiceId": "2103.в.27",
+      "counterEdrpou": "3096889974",
+      "counterIban": "UA898999980000355639201001404"
+    }
+  }
+}
+  ''';
+
+  // Parsing webhook data
+  final webhookEvent =
+      WebhookEvent.fromJson(jsonDecode(webhookData) as Map<String, dynamic>);
+
+  // Checking webhook type
+  if (webhookEvent is StatementItemWebhookEvent) {
+    // Printing raw statement item
+    print('${webhookEvent.item.mcc.emoji} ${webhookEvent.item}');
+
+    // Fetching account data for proper data
+    final client = MonoAPI('token');
+    await webhookEvent.account.resolve(client);
+    final newItem = webhookEvent.item.regenerate();
+
+    // Printing regenerated statement item
+    //
+    // Here you can see proper cashback type, since this data relies on
+    // account data.
+    //
+    // Also in this example amount == operationAmount, so StatementItem
+    // assumes currencyCode is the same as currencyCode of the account.
+    //
+    // This may be not true in real life, and in such cases you will have
+    // [Currency.dummy] currency.
+    print('${newItem.mcc.emoji} $newItem');
+  }
 }